+++ src/Lunil.CodeGen.Cil/Jit/ReflectionEmitLuaNumericRegionCompiler.cs
@@ -249,15 +249,15 @@ internal static class ReflectionEmitLuaNumericRegionCompiler
     private static readonly MethodInfo ContinueExit = Method(
         typeof(LuaCompiledExit),
         nameof(LuaCompiledExit.Continue),
-        [typeof(int), typeof(int)]);
+        [typeof(int), typeof(long)]);
     private static readonly MethodInfo PollExit = Method(
         typeof(LuaCompiledExit),
         nameof(LuaCompiledExit.Poll),
-        [typeof(int), typeof(int), typeof(LuaCompiledExitReason)]);
+        [typeof(int), typeof(long), typeof(LuaCompiledExitReason)]);
     private static readonly MethodInfo DeoptExit = Method(
         typeof(LuaCompiledExit),
         nameof(LuaCompiledExit.Deopt),
-        [typeof(int), typeof(int), typeof(LuaCompiledExitReason)]);
+        [typeof(int), typeof(long), typeof(LuaCompiledExitReason)]);
     private static readonly ConstructorInfo InvalidOperationExceptionConstructor =
         typeof(InvalidOperationException).GetConstructor([typeof(string)]) ??
         throw new MissingMethodException(
